% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{as_month}
\alias{as_month}
\alias{as_week}
\title{Extract the calendar period components from date-times}
\usage{
as_month(x)

as_week(x)
}
\arguments{
\item{x}{Date-times.}
}
\value{
The period of \code{x} as date format or character string. It always
   returns the first day of the month (\code{as_month}) and the first day of
   the week (\code{as_week}).
}
\description{
Extract the higher-level calendar period components from date-times
}
\examples{
   x <- as.POSIXct(c("2014-01-31", "2015-07-31", "2016-10-31"))
   as_month(x)
   as_week(x)

}
\author{
Earo Wang
}

