% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{as_tsibble}
\alias{as_tsibble}
\title{Coerce to a tsibble object}
\usage{
as_tsibble(x, key = key_vars(), index, ...)
}
\arguments{
\item{x}{Other objects to be coerced to tsibble.}

\item{key}{Unquoted variable(s) indicating the key variables for tsibble,
used in combination with \code{key_vars()}.}

\item{index}{An unquoted variable indicating the time index variable}

\item{...}{Other arguments to be passed.}
}
\value{
A tsibble object.
}
\description{
Coerce to a tsibble object
}
\examples{
   tidypkgs_ts <- as_tsibble(tidypkgs, key = key_vars(package), index = date) 
   print(tidypkgs_ts)

}
\seealso{
\link[tibble:as_tibble]{tibble::as_tibble}
}
\author{
Earo Wang
}
