% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_period.R
\name{by_month}
\alias{by_day}
\alias{by_month}
\alias{by_week}
\title{Aggregate over fixed calendar periods}
\usage{
by_month(.data, .index, .value, .f, ..., .label = ".out")

by_week(.data, .index, .value, .f, ..., .label = ".out")

by_day(.data, .index, .value, .f, ..., .label = ".out")
}
\arguments{
\item{.data}{A data frame.}

\item{.index}{A variable of date-time objects in the \code{.data}.}

\item{.value}{A variable of numerics in the \code{.data} to be aggregated.}

\item{.f}{A function applied to aggregation over the fixed calendar periods.}

\item{...}{The extra arguments passed to \code{.f}.}

\item{.label}{a character string to be labelled for the aggregated values.}
}
\value{
A data frame that contains the calendar periods and the aggregated values.
}
\description{
Aggregate over calendar periods
}
\examples{
   library(dplyr)
   pedestrian \%>\%
     group_by(Sensor_Name, Sensor_ID) \%>\% 
     by_month(.index = Date_Time, .value = Hourly_Counts, .f = quantile)

}
\author{
Earo Wang
}

