% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key.R
\name{draw_key_acf}
\alias{draw_key_acf}
\title{Key drawing functions}
\usage{
draw_key_acf(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob
}
\description{
Key drawing functions
}
