% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-calendar.R
\docType{data}
\name{facet_calendar}
\alias{facet_calendar}
\alias{FacetCalendarMday}
\title{Lay out panels in a calendar}
\usage{
facet_calendar(date, nrow = NULL, ncol = NULL,
  calendar.label = "mday", scales = "fixed", shrink = TRUE,
  strip.position = "top")
}
\arguments{
\item{date}{A variable that contains "Date" class.}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}

\item{calendar.label}{Either "wday" or "mday".}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
  "bottom", "left", "right")}}
}
\description{
Lay out panels in a calendar
}
\examples{
fs <- pedestrian \%>\%
  dplyr::filter(Sensor_Name == "Flagstaff Station", Date < as.Date("2016-05-01"))

fs \%>\%
  ggplot(aes(x = Time, y = Hourly_Counts, group = Date)) +
  geom_line() +
  facet_calendar(date = Date, nrow = 2)
}
\keyword{datasets}
