% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{filter.tbl_ts}
\alias{filter.tbl_ts}
\title{Return rows with matching conditions}
\usage{
\method{filter}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}}

\item{...}{Logical expression.}
}
\value{
A tsibble
}
\description{
Return rows with matching conditions
}
\examples{
   # Leave blank
}
\seealso{
\link[dplyr:filter]{dplyr::filter}
}
\author{
Earo Wang
}
