% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{group_by.tbl_ts}
\alias{group_by.tbl_ts}
\title{Group by one or more variabels}
\usage{
\method{group_by}{tbl_ts}(.data, ..., add = FALSE)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{Unquoted variables to be grouped by.}

\item{add}{Logical. \code{FALSE} is the default suggesting to override the exisiting
groups. \code{TRUE} for adding the exisiting groups.}
}
\value{
A grouped tsibble
}
\description{
Group by one or more variabels
}
\examples{
   pkgs_ts <- as_tsibble(tidypkgs, key = key_vars(package), index = date)
   group_by(pkgs_ts, package)

}
\seealso{
\link[dplyr:group_by]{dplyr::group_by}
}
\author{
Earo Wang
}
