% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{mutate.tbl_ts}
\alias{mutate.tbl_ts}
\title{Add new variables}
\usage{
\method{mutate}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{Name-value pairs of expressions.}
}
\value{
A tsibble.
}
\description{
Add new variables
}
\examples{
   pkgs_ts <- as_tsibble(tidypkgs, key = key_vars(package), index = date)
   pkgs_ts \%>\% 
     group_by(package) \%>\% 
     mutate(pkg_avg = mean(count))
   
}
\seealso{
\link[dplyr:mutate]{dplyr::mutate}
}
\author{
Earo Wang
}
