% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedestrian}
\alias{pedestrian}
\title{Pedestrian counts in Melbourne city}
\format{A data frame with 70072 rows and 4 variables:
\describe{
  \item{Date_Time}{Date time in 2015}
  \item{Sensor_ID}{Sensor identifiers}
  \item{Sensor_Name}{Sensor names}
  \item{Hourly_Counts}{Pedestrian Counts at hourly intervals}
}}
\usage{
pedestrian
}
\description{
A dataset containing the pedestrian counts at hourly intervals over the year
of 2015 at 8 sensors in the city of Melbourne. The variables are as follows:
}
\examples{
pedestrian
}

