% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3.R
\name{d3_sunburst}
\alias{d3_sunburst}
\alias{d3_indented_tree}
\title{Plot D3 hierarchy}
\usage{
d3_sunburst(data, width = NULL, height = NULL)

d3_indented_tree(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{A parent-name paired data frame}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}
}
\description{
Plot D3 hierarchy
}
