% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{vars}
\alias{reexports}
\alias{year}
\alias{reexports}
\alias{as_date}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{vars}}}

  \item{lubridate}{\code{\link[lubridate]{year}}, \code{\link[lubridate]{as_date}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

