% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{season}
\alias{season}
\title{Get seasons component of a date-time}
\usage{
season(x, hemisphere = c("N", "S"), definition = c("meterological",
  "astronomical", "midsolstice"))
}
\arguments{
\item{x}{A date-time object.}

\item{hemisphere}{A character vector that specifies the hemisphere. "N" refers
to the Northern hemisphere, and "S" to the Souther hemisphere.}

\item{definition}{A character vector that defines the season range. "meterological"
is the only one supported at the stage.}
}
\value{
The season of \code{x} as character string.
}
\description{
Get seasons component of a date-time
}
\examples{
   x <- as.POSIXct(c("2014-01-31", "2015-07-31", "2016-10-31"))
   season(x, hemisphere = "S") # I'm in Australia

}
\author{
Earo Wang
}
\references{
Barnett, A. G. and Dobson A. J. (2010) Analysing seasonal health data. Springer.
}
\seealso{
as_season
}

