% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{select.tbl_ts}
\alias{select.tbl_ts}
\title{Select variables by name}
\usage{
\method{select}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_ts}.}

\item{...}{Unquoted variables to be selected.}
}
\value{
A tsibble when keeping the key and the index variables.
}
\description{
Select variables by name
}
\examples{
   # Leave blank
}
\seealso{
\link[dplyr:select]{dplyr::select}
}
\author{
Earo Wang
}
