% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Rearrange data into parent-children pair data for d3-hierarchy}
\usage{
stratify(data, ..., size = 1L, root, na.rm = FALSE)
}
\arguments{
\item{data}{A tabular data.}

\item{...}{Nesting variables.}

\item{size}{\itemize{
\item \code{1} means identity
\item Bare variable name
\item A vector of the same length as the distinct key
}}

\item{root}{Character.}

\item{na.rm}{\code{TRUE} removes \code{NA} in the key.}
}
\description{
Rearrange data into parent-children pair data for d3-hierarchy
}
