% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugrrants-pkg.R
\docType{package}
\name{sugrrants-package}
\alias{sugrrants-package}
\title{sugrrants: supporting graphs for analysing time series}
\description{
Provides 'ggplot2' graphics for analysing time series data. It aims to fit
into the 'tidyverse' and grammar of graphics framework for handling temporal
data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.earo.me/sugrrants}
  \item Report bugs at \url{https://github.com/earowang/sugrrants/issues}
}

}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com} (0000-0001-6448-5260)

Authors:
\itemize{
  \item Di Cook (0000-0002-3813-7155) [thesis advisor]
  \item Rob Hyndman (0000-0002-2140-5352) [thesis advisor]
}

}
