% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sweden_temp}
\alias{sweden_temp}
\title{Daily temperature recorded at 17 stations across Sweden}
\format{A data frame with 64979 rows and 5 variables:
\describe{
  \item{lon}{Longitude of the weather stations}
  \item{lat}{Latitude of the weather stations}
  \item{date}{Recorded dates}
  \item{temp}{Daily temperature in Celsius degree}
  \item{locations}{Location index}
}}
\usage{
sweden_temp
}
\description{
A dataset containing daily temperatures from 1998 to 2008 at 17 weather stations
across Sweden. The variables are as follows:
}
\examples{
sweden_temp
}

