% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidypkgs}
\alias{tidypkgs}
\title{Tidyverse core packages daily downloads}
\format{A data frame with 4386 rows and 3 variables:
\describe{
\item{date}{Date when the downloads are made}
\item{count}{Daily downloads}
\item{package}{Tidyverse core package name}
}}
\usage{
tidypkgs
}
\description{
A dataset containing the daily downloads of tidyverse core packages from
2015-01-01 to 2016-12-31. The variables are as follows:
}
\examples{
   tidypkgs
}
\keyword{datasets}
