% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{wday2}
\alias{wday2}
\title{Get weekdays component of a date-time}
\usage{
wday2(x, label = TRUE, start = "Mon")
}
\arguments{
\item{x}{A date-time object.}

\item{label}{Logical. If TRUE, it displays the day of the week as an ordered
character strings.}

\item{start}{A vector of character. Only supports the abbreviations of the
weekday.}
}
\value{
The day of the week
}
\description{
A wrapper of \code{lubridate::wday()}, but allows to define the
   starting weekday.
}
\examples{
   x <- as.POSIXct(c("2014-01-31", "2015-07-31", "2016-10-31"))
   wday2(x)

}
\author{
Earo Wang
}
\seealso{
\link[lubridate]{wday}
}

