% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{as_season}
\alias{as_month}
\alias{as_season}
\alias{as_week}
\title{Extract the calendar period components from a date-time}
\usage{
as_season(x, hemisphere = "N", definition = "meterological",
  date_format = TRUE)

as_month(x, date_format = TRUE)

as_week(x, date_format = TRUE)
}
\arguments{
\item{x}{A date-time object.}

\item{hemisphere}{A character vector that specifies the hemisphere. "N" refers
to the Northern hemisphere, and "S" to the Souther hemisphere.}

\item{definition}{A character vector that defines the season range. "meterological"
is the only one supported at the stage.}

\item{date_format}{Logical. If TRUE, it returns a date format object; otherwise
a character string.}
}
\value{
The period of \code{x} as date format or character string.
}
\description{
Extract the calendar period components from a date-time
}
\examples{
   x <- as.POSIXct(c("2014-01-31", "2015-07-31", "2016-10-31"))
   as_month(x)

}
\author{
Earo Wang
}
\references{
Barnett, A. G. and Dobson A. J. (2010) Analysing seasonal health data. Springer.
}

