% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_period.R
\name{by_month}
\alias{by_month}
\alias{by_week}
\alias{by_week}
\alias{by_day}
\title{Aggregate over fixed calendar periods}
\usage{
by_month(.data, .index, .value, .f, ..., .label = ".out")

by_week(.data, .index, .value, .f, ..., .label = ".out")

by_day(.data, .index, .value, .f, ..., .label = ".out")
}
\arguments{
\item{.data}{A data frame.}

\item{.index}{A variable of date-times in the \code{.data}.}

\item{.value}{A variable in the \code{.data} that a specified function to
apply.}

\item{.f}{A specified function applied to the fixed calendar windows of the
\code{.value}.}

\item{...}{The extra arguments passed to \code{.f}.}

\item{.label}{a character string indicates the name of output column.}
}
\value{
An aggregated data frame with the added variables of calendar periods 
   (e.g. \code{.month}, \code{.week}, \code{.day}) and the aggregated values 
   while keeping grouped variables if present.
}
\description{
Apply a specified function to each fixed calendar periods of
   a data frame that contains a date-times variable. It can be used in
   conjunction with \code{dplyr::group_by}.
}
\details{
When a specified function \code{.f} generates more than one values,
   it returns a long data form with one stack on the rest.
}
\examples{
   library(dplyr)
   pedestrian \%>\%
     group_by(Sensor_Name) \%>\% 
     by_month(.index = Date_Time, .value = Hourly_Counts, .f = quantile,
       na.rm = TRUE)

}
\author{
Earo Wang
}
