% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_period.R
\name{by_season}
\alias{by_month}
\alias{by_season}
\alias{by_week}
\title{Aggregate over fixed calendar periods}
\usage{
by_season(.data, .index, .value, .f, ..., .label = ".out",
  hemisphere = c("N", "S"), definition = c("meterological", "astronomical",
  "midsolstice"), date_format = TRUE)

by_month(.data, .index, .value, .f, ..., .label = ".out",
  date_format = TRUE)

by_week(.data, .index, .value, .f, ..., .label = ".out", date_format = TRUE)
}
\arguments{
\item{.data}{A data frame.}

\item{.index}{A variable of date-time objects in the \code{.data}.}

\item{.value}{A variable of numerics in the \code{.data} to be aggregated.}

\item{.f}{A function applied to aggregation over the fixed calendar periods.}

\item{...}{The extra arguments passed to \code{.f}.}

\item{.label}{a character string to be labelled for the aggregated values.}

\item{hemisphere}{A character vector that specifies the hemisphere. "N" refers
to the Northern hemisphere, and "S" to the Souther hemisphere.}

\item{definition}{A character vector that defines the season range. "meterological"
is the only one supported at the stage.}

\item{date_format}{Logical. If TRUE, it returns a date format; otherwise a
character string.}
}
\value{
A data frame that contains the calendar periods and the aggregated values.
}
\description{
Aggregate over calendar periods
}
\author{
Earo Wang
}

