% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-calendar.R
\docType{data}
\name{facet_calendar}
\alias{facet_calendar}
\alias{FacetCalendar}
\title{Lay out panels in a calendar}
\usage{
facet_calendar(date, nrow = NULL, ncol = NULL, format = "\%b \%d",
  week_start = getOption("lubridate.week.start", 1), scales = "fixed",
  shrink = TRUE, dir = "h", strip.position = "top")
}
\arguments{
\item{date}{A variable that contains "Date" class.}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}

\item{format}{A character string. See \code{?strptime} for details.}

\item{week_start}{Day on which week starts following ISO conventions -
1 means Monday, 7 means Sunday (default). You can set \code{lubridate.week.start}
option to control this parameter globally.}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{dir}{Direction: either \code{"h"} for horizontal, the default, or \code{"v"},
for vertical.}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
  "bottom", "left", "right")}}
}
\description{
Lay out panels in a calendar
}
\examples{
fs <- pedestrian \%>\%
  dplyr::filter(Sensor_Name == "Flagstaff Station", Date < as.Date("2016-05-01"))

fs \%>\%
  ggplot(aes(x = Time, y = Hourly_Counts, group = Date)) +
  geom_line() +
  facet_calendar(date = Date, nrow = 2)
}
\keyword{datasets}
