% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_calendar.R
\name{frame_calendar}
\alias{frame_calendar}
\title{Prepare a data frame for a ribbon plot using \code{ggplot2}}
\usage{
frame_calendar(.data, x, y, date, nrow = NULL, ncol = NULL)
}
\arguments{
\item{.data}{A data frame.}

\item{x, y}{A variable to be passed to \code{ggplot2(aes(x, y))}.}

\item{date}{A variable that defines the calendar.}

\item{nrow, ncol}{Number of rows and columns.}
}
\value{
A data frame consisting of new \code{.x}, \code{.y}, and \code{.group_id}
   columns.
}
\description{
Prepare a data frame for a ribbon plot using \code{ggplot2}
}
\examples{
   library(dplyr); library(lubridate)
   pedestrian \%>\%
     filter(Sensor_ID == 13) \%>\% 
     mutate(Time = hour(Date_Time)) \%>\% 
     frame_calendar(x = Time, y = Hourly_Counts, date = Date_Time,
       nrow = 3, ncol = 4)

}
\author{
Earo Wang
}

