% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_geom.R
\name{frame_ribbon}
\alias{frame_ribbon}
\title{Prepare a data frame for a ribbon plot using \code{ggplot2}}
\usage{
frame_ribbon(.data, .value)
}
\arguments{
\item{.data}{A data frame in a long form}

\item{.value}{A variable containing numerics in order to produce the ribbon.}
}
\value{
A data frame consisting of new \code{lower}, \code{upper}, possiblly
   \code{middle} and \code{pair} columns.
}
\description{
Prepare a data frame for a ribbon plot using \code{ggplot2}
}
\examples{
   library(dplyr)
   pedestrian \%>\%
     group_by(Sensor_Name, Sensor_ID) \%>\% 
     by_month(.index = Date_Time, .value = Hourly_Counts, .f = quantile) \%>\% 
     frame_ribbon(.value = .out)

}
\author{
Earo Wang
}

