% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_ribbon.R
\name{frame_ribbon}
\alias{frame_ribbon}
\title{Get a data frame of long form ready for \code{ggplot2::geom_ribbon}}
\usage{
frame_ribbon(.data, .value)
}
\arguments{
\item{.data}{A grouped data frame in a long form in conjunction with \code{group_by}.}

\item{.value}{A variable of numerics in order to produce the ribbon bounds.}
}
\value{
A data frame with new columns of \code{lower}, \code{upper}, possibly
   \code{middle} and \code{pair}.
}
\description{
\code{ggplot2::geom_ribbon} requires at least two aesthetics mapping of 
   \code{ymin} and \code{ymax} to make a ribbon plot. \code{frame_ribbon} lubricates
   the transition from a long data form to the basic data form that is wanted for
   \code{ggplot2::geom_ribbon}.
}
\details{
A ribbon plot can be seen as a simplified version of boxplot. However,
   the ribbon plot employs a smoother view if the continuous variable is used
   as one of the axes compared to the boxplot does for the categorical one.
}
\examples{
   # convert the summarised data frame to the ribbon format
   library(dplyr)
   ped_ribbon <- pedestrian \%>\%
     group_by(Sensor_Name, Sensor_ID) \%>\% 
     by_month(.index = Date_Time, .value = Hourly_Counts, .f = quantile) \%>\% 
     frame_ribbon(.value = .out)
   
   # use ggplot2 to produce the ribbon plot
   library(ggplot2)
     ggplot(ped_ribbon, aes(x = .month)) +
     geom_ribbon(aes(ymin = lower, ymax = upper, 
       alpha = .pair, colour = Sensor_Name)) +
     geom_line(aes(y = middle)) +
     facet_grid(Sensor_Name ~ ., scales = "free_y")

}
\author{
Earo Wang
}

