% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Rearrange data into parent-children pair data for d3-hierarchy}
\usage{
stratify(data, formula, root, na.rm = FALSE)
}
\arguments{
\item{data}{A tabular data.}

\item{formula}{Node size on the lhs, nesting variables on the rhs.}

\item{root}{Character.}

\item{na.rm}{\code{TRUE} removes \code{NA} in the key.}
}
\description{
Rearrange data into parent-children pair data for d3-hierarchy
}
