% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{wday2}
\alias{wday2}
\title{Get weekdays component of a date-time}
\usage{
wday2(x, label = FALSE, abbr = TRUE)
}
\arguments{
\item{x}{A date-time object.}

\item{label}{Logical. If TRUE, it displays the day of the week as an ordered
character strings; otherwise, an integer returns}

\item{abbr}{Logical. If TRUE, it returns an abbreviated version the label.}
}
\value{
The day of the week as a number (Monday is 1) or an ordered factor
   (Monday is first).
}
\description{
A wrapper of \code{lubridate::wday()}, but allows to define the
   starting weekday.
}
\examples{
   x <- as.POSIXct(c("2014-01-31", "2015-07-31", "2016-10-31"))
   wday2(x, label = TRUE)

}
\author{
Earo Wang
}
\seealso{
\link[lubridate]{wday}
}

